Preparing 2.10
git-svn-id: https://stealth.svn.sourceforge.net/svnroot/yodl/trunk@96 7de777c3-3b1c-0410-83c5-80958bd9f4b1

+++ yodl/src/chartab/ctstring.c
@@ -11,7 +11,7 @@ char *ct_string(char const *str)
         string_addchar
         (
             &string,
-            (*str == '\\') ? ct_ascii(str++) : *str
+            (*str == '\\') ? (int)ct_ascii(str++) : *str
         );
         str++;
     }
