+++ yodl/src/yodlpost/handlexmldone.c
@@ -4,7 +4,7 @@ void handle_xml_done(long offset, HashItem *item)
 {
     file_copy2offset(global.d_out, postqueue_istream(), offset);
 
-    xml_chapterlinks(global.d_fileNr++);
+    xml_chapterlinks((int)global.d_fileNr++);
 
     file_append(global.d_out, hashmap_textOf(&global.d_symbol,
                                                         "XXxhtmlEnd"));
